(feat) flatten events and command directories
